ast/compile: also replace declared vars in with's target (#6980)
It might be uncommon to do this, but it's not wrong to expect this to work.
One peculiarity is that we've consciously made an exception for
```
input := "whatever"
allow with input as true
```
where the with-target "input" was meant not to be replaced although there's
a local variable shadowing "input". This change has become a bit trickier
because this behaviour is kept intact. Note that it's only possible in non-strict
mode (Rego v0).

Fixes #6979

Small refactoring included:
* ast/compile: remove With from recursive term rewriting

   I could find no term where this actually mattered, so I've simplified the code
   a little bit.

Signed-off-by: Stephan Renatus <stephan@styra.com>