debug: Always including `Input` and `Data` variable scopes
even when there are no explicit `input` and `data` documents.

This makes discoverability of the scopes much better, as users will know to expect them.
Whereas, if it's only enabled/visible when there is an actual document, a user who
has not come across it before might not realize it should appear, and can therefore
be slow to realize scenarios where their setup is expecting an `input` document, but
is missing it for some reason.

Signed-off-by: Johan Fylling <johan.dev@fylling.se>