test: Parallelize package level tests in high-cost packages.
This commit adds `t.Parallel()` calls to the beginning of many tests
across several Go packages in OPA. The slowest packages (taking ~10s or
more) have been instrumented where possible as a proof-of-concept. On a
machine with many cores, the tests now will complete as fast as the
slowest test per package, instead of the sum of all the tests in a
particular package.

Background: Go runs tests across different packages in parallel, but
will run all tests *within* a package sequentially. By adding
`t.Parallel()` to tests that support it, we are telling the test-runner
that it is safe to run that test in parallel with any other tests.

Signed-off-by: Philip Conrad <philip@chariot-chaser.net>