topdown: move context.Context cancellation check
The topdown Cancel machinery is there because it's cheap to check. ctx.Err() is
not.

This change moves the "Is the context the cause for cancellation?" check into
the branch were evaluation has been aborted through the topdown.Cancel call.

When evaluation has already been cancelled, an expensive check no longer
matters much -- when it's still ongoing, it'll affect the overall performance.

Signed-off-by: Stephan Renatus <stephan@styra.com>