eval: optimize iteration (#7327)
Since we do a lot of this in Rego, I was curious to learn what the cost
of that was, and if it was possible to do better. I identified the two
main costs as these:

- `handleErr` wrapper allocating in each iteration, even though the
  normal condition is to not handle an error
- `biunify` callback function literal leaking to the heap in each
  iteration, with an _high_ cost associated for such a simple one-liner
  function. While it wasn't possible to change that (as it referenced
  the outer scope) it *was* possible to inline the whole biunify flow
  for the array iteration case, as that turned out to be quite simple
  when stepped through with a debugger.

Ideally we'd be able to inline iteation over sets and objects too, but
that's much more complex (see comment in code), so I left that for the
future.

The numbers benchmarked are in the PR, so take a look there. The array
iteration case is *quite* the improvement!

For reference, this amounts to about 2 million allocations less in the
`regal lint bundle` benchmark, and a noticeable improvement in
evaluation time.

Signed-off-by: Anders Eknert <anders@styra.com>