MAINTAINERS: Add Anders and Charlie as maintainers (#7318)
This commit adds Anders and Charlie as maintainers under the `opa` area
of expertise. This reflects active contributions to the project and
aids in managing the project's relationship with the CNCF as maintainer
status is required for various KubeCon related activities.

Since this does not add any new organizations, it does not impact the
project governance.

Signed-off-by: Charlie Egan <charlie@styra.com>