Make oracle public under v1/ast/oracle (#7352)
I'll follow up with another PR to modify this to allow passing
a custom compiler, and whatever else we need in Regal. But moving
files *and* modifying them in the same change is rarely great for
reviewing. So this does nothing but move the package and adjust
the pointer to the package in cmd/oracle.

Fixes #7265

Signed-off-by: Anders Eknert <anders@styra.com>