perf: StringTerm interning
Allow interning of String values and terms, and have functions
like `ast.InterfaceToValue` leverage that capability when possible.

This first PR mainly impacts the store when configured to store
AST objects, as in that case all string keys are interned. But any
function calling `InterfaceToValue` will benefit, provided that the
term is interned since before.

Memory consumption for the bundle I used for testing went from
800 MB to 500 MB.

Signed-off-by: Anders Eknert <anders@styra.com>