Delete reference to license key in envoy-tutorial-standalone-envoy.md
Apparently setting a license key is not (longer?) needed. The tutorial doesn't mention it in the rest of the text as the deleted line promises. I couldn't find a hidden statement about a license key in the config files. Not does the page on installing opa using docker mention a license key. https://www.openpolicyagent.org/docs/latest/deployments/

Signed-off-by: Joost Holslag <joostholslag@users.noreply.github.com>