fix: don't drop adaptive uncompressed size limit on upload and drop ND cache sparingly
Renamed the "soft" limit to "uncompressed limit" throughout the code and documentation for clarity.
In the size and event buffer the uncompressed limit was being dropped after each upload, now it is carried over. The event buffer doesn't reset the encoder at all. Checking if an individual size is too big was comparing the uncompressed limit to the compressed limit causing events to be dropped or lose the ND cache unnecesarily. This is now fixed, instead if the uncompressed limit allows it the event is compressed and then multiple attempts are made before losing the ND cache or dropping the event. The configurable upload is used to calculate the uncompressed size by exponentially growing it, this could cause an overflow if it was set too high. Added a max.

Signed-off-by: sspaink <sspaink@styra.com>