refactor: Remove internal/gqlparser and use upstream instead.
Upstream has incorporated all the changes OPA needs, so maintaining an internal
version is no longer necessary.

Two kinds of changes were required to GraphQL Test Cases for compatibility with
the current version of the gqlparser library:

1. Update the input to be valid GraphQL according to the reference implementation.
   Previously, the GraphQL parsing library would automatically add in required fields
   that may be missing.
2. Removed comments.  Previously, the GraphQL parsing library did not include comments
   in the generated ASTs.

This updates gqlparser from version ~v2.5.1 and the June 2018 GraphQL spec to
v2.5.26 and the October 2021 GraphQL spec.

Signed-off-by: Rob Myers <1243316+robmyersrobmyers@users.noreply.github.com>