docs: Remove live blocks
These are no longer being used and are dependent on an old version of
node to work.

I have disabled the generation to allow the new docs site to be built. I
thought it made sense to remove all things relating to live blocks since
we have no plans to support them now anyway.

Signed-off-by: Charlie Egan <charlie@styra.com>