docs: Allow searching of ecosystem page (#7574)
* docs: Allow searching of ecosystem page

At the moment, showing the long list is not ideal without categorizing
the items. However, search helps a little as well as the existing categorization
earlier in the page.

This PR also adds a small note to show how the page is ordered.

Signed-off-by: Charlie Egan <charlie@styra.com>

* docs: No results text for ecosystem search

Links to slack to in case the user is confused.

Signed-off-by: Charlie Egan <charlie@styra.com>

---------

Signed-off-by: Charlie Egan <charlie@styra.com>