docs: Reliably link to builtin anchors (#7619)
Old links used no . in the anchor name. This has been restored. We also
have an issue where the page is to large it takes too long for the anchor
to be available and so the browser does not jump to the row. This page
is ideally to be broken down soon, but for now we have some JS code that
runs to point the user to the right location.

Signed-off-by: Charlie Egan <charlie@styra.com>