docs: Revise card display across the site (#7634)
This was mostly an issue on /ecosystem but also affected any pages with
many cards. Cards are now rendered in a grid layout in the same way
across the site. Some minor updates to content land here too to make the
update to the page look better above the fold.

Signed-off-by: Charlie Egan <charlie@styra.com>