Fix issue where path in `walk` would get mutated
I'm still not sure where this mutation occurs, as the append operation
done in `walk` returns a copy and shouldn't mutate. But at least creating
a shallow copy rather than a slice fixes the issue, and is still quite
performant compared to the previous implementation that did a deep-copy
of the path in each iteration.

Fixes #7656

Signed-off-by: Anders Eknert <anders@styra.com>