v1/metrics: Fix restartable timer bug.
This commit fixes a potential bug that could occur with
`metrics.Timer`. Whenever `timer.Stop()` was called, the
timer would add the current delta from the start time to its
accumulated value so far, *regardless of whether it had been
restarted with `timer.Start()` or not*.

This could lead to confusing metrics in some places if `timer.Stop()`
was accidentally called twice. (Value would be much larger than
expected.)

The fix is to use a guard around the accumulating step.

Co-authored-by: Anders Eknert <anders@styra.com>
Signed-off-by: Philip Conrad <philip@chariot-chaser.net>