docs: Add EvergreenCodeBlock for code with version (#7706)
This adds a new EvergreenCodeBlock component to the documentation site that allows
code blocks to contain template for the current versions of opa
components. This means we can show the current version in docs, but
don't need update the version in every code block when we release a new
version.

Fixes https://github.com/open-policy-agent/opa/issues/7705

In the future, it would be nice to pull the latest release for this on
build since the envoy release sometimes lags behind the opa release a
small amount, but for now this is a step in the right direction. The
kube-mgmt version is also manually set, but at least this component owns
all versioned code blocks in the documentation now.

Signed-off-by: Charlie Egan <charlie@styra.com>