docs: Show TOC on CLI page (#7712)
* docs: CLI show TOC

This page has no table of contents since the contents is generated in
the command list component.

This exports `toc` which is docusaurus magic to set the table of contents
values manually. The command ids are used as these are also used as
anchor names in CommandList.

Signed-off-by: Charlie Egan <charlie@styra.com>

* docs: CommandDoc correct anchor jump location

Also export styles to styles.module.css

Signed-off-by: Charlie Egan <charlie@styra.com>

---------

Signed-off-by: Charlie Egan <charlie@styra.com>