docs: Show edge rather than latest release (#7717)
* docs: Show edge rather than latest release

Docs are updated as we push to main, this means the version is not the
latest, but rather the edge version. This is to avoid confusion while
still showing fixes to the docs site as we make them, since it's still
an active area of development after we released the new version.

Signed-off-by: Charlie Egan <charlie@styra.com>

* docs: Allow setting of builtVersion

This will be used to set the version in the navbar item on versioned
deploys of the site.

Signed-off-by: Charlie Egan <charlie@styra.com>

---------

Signed-off-by: Charlie Egan <charlie@styra.com>