docs: Update sidebar (#7723)
I have had some feedback that the sidebar is too long and disorganized.
The sidebar has not been updated much in the transition to the new docs site.

I am expecting some feedback here, but this is a first pass.

The docs archive has been removed as this is now linked from the top of
the docs nav.

Signed-off-by: Charlie Egan <charlie@styra.com>