perf: Only pass built-in context to calls depending on it (#7728)
This commit updates the ast.Builtin struct adding a private field
that'll store information about whether the built-in needs a builtin
context, and a public method to read this.

This information is then read in eval, where the construction of
the BuiltinContext is skipped unless a built-in depends on that.

Future work will allow custom built-ins to register this as well.

Signed-off-by: Anders Eknert <anders@styra.com>